+++ lightly_studio/src/lightly_studio/resolvers/video_frame_resolver/get_all_by_video_ids.py
@@ -7,6 +7,7 @@ from uuid import UUID
 
 from sqlmodel import Session, col, select
 
+from lightly_studio import db_array
 from lightly_studio.models.video import VideoFrameTable
 
 
@@ -23,7 +24,7 @@ def get_all_by_video_ids(
 
     stmt = (
         select(VideoFrameTable)
-        .where(col(VideoFrameTable.parent_sample_id).in_(list(video_ids)))
+        .where(db_array.in_array(column=col(VideoFrameTable.parent_sample_id), values=video_ids))
         .order_by(
             col(VideoFrameTable.parent_sample_id).asc(), col(VideoFrameTable.frame_number).asc()
         )